storage: move some gc stuff out of reader
