+++ src/storage/mvcc/reader/mod.rs
@@ -5,9 +5,9 @@ mod reader;
 mod scanner;
 
 pub use self::point_getter::{PointGetter, PointGetterBuilder};
-pub use self::reader::{
-    check_need_gc, MvccReader, OverlappedWrite, SnapshotReader, TxnCommitRecord,
-};
+#[cfg(test)]
+pub use self::reader::tests as reader_tests;
+pub use self::reader::{MvccReader, OverlappedWrite, SnapshotReader, TxnCommitRecord};
 pub use self::scanner::test_util;
 pub use self::scanner::{
     has_data_in_range, seek_for_valid_write, DeltaScanner, EntryScanner, Scanner, ScannerBuilder,
