+++ src/storage/mvcc/reader/reader.rs
@@ -2,7 +2,7 @@
 
 use crate::storage::kv::{Cursor, CursorBuilder, ScanMode, Snapshot, Statistics};
 use crate::storage::mvcc::{default_not_found_error, seek_for_valid_write, Result};
-use engine_traits::{IterOptions, MvccProperties};
+use engine_traits::IterOptions;
 use engine_traits::{CF_DEFAULT, CF_LOCK, CF_WRITE};
 use txn_types::{Key, Lock, OldValue, TimeStamp, Value, Write, WriteRef, WriteType};
 
@@ -537,36 +537,8 @@ impl<S: Snapshot> MvccReader<S> {
     }
 }
 
-// Returns true if it needs gc.
-// This is for optimization purpose, does not mean to be accurate.
-pub fn check_need_gc(safe_point: TimeStamp, ratio_threshold: f64, props: &MvccProperties) -> bool {
-    // Always GC.
-    if ratio_threshold < 1.0 {
-        return true;
-    }
-
-    // No data older than safe_point to GC.
-    if props.min_ts > safe_point {
-        return false;
-    }
-
-    // Note: Since the properties are file-based, it can be false positive.
-    // For example, multiple files can have a different version of the same row.
-
-    // A lot of MVCC versions to GC.
-    if props.num_versions as f64 > props.num_rows as f64 * ratio_threshold {
-        return true;
-    }
-    // A lot of non-effective MVCC versions to GC.
-    if props.num_versions as f64 > props.num_puts as f64 * ratio_threshold {
-        return true;
-    }
-
-    false
-}
-
 #[cfg(test)]
-mod tests {
+pub mod tests {
     use super::*;
     use crate::storage::kv::Modify;
     use crate::storage::mvcc::{tests::write, MvccReader, MvccTxn};
@@ -591,20 +563,20 @@ mod tests {
     use std::u64;
     use txn_types::{LockType, Mutation};
 
-    struct RegionEngine {
+    pub struct RegionEngine {
         db: Arc<DB>,
         region: Region,
     }
 
     impl RegionEngine {
-        fn new(db: &Arc<DB>, region: &Region) -> RegionEngine {
+        pub fn new(db: &Arc<DB>, region: &Region) -> RegionEngine {
             RegionEngine {
                 db: Arc::clone(&db),
                 region: region.clone(),
             }
         }
 
-        fn put(
+        pub fn put(
             &mut self,
             pk: &[u8],
             start_ts: impl Into<TimeStamp>,
@@ -616,7 +588,7 @@ mod tests {
             self.commit(pk, start_ts, commit_ts);
         }
 
-        fn lock(
+        pub fn lock(
             &mut self,
             pk: &[u8],
             start_ts: impl Into<TimeStamp>,
@@ -628,7 +600,7 @@ mod tests {
             self.commit(pk, start_ts, commit_ts);
         }
 
-        fn delete(
+        pub fn delete(
             &mut self,
             pk: &[u8],
             start_ts: impl Into<TimeStamp>,
@@ -640,7 +612,7 @@ mod tests {
             self.commit(pk, start_ts, commit_ts);
         }
 
-        fn txn_props(
+        pub fn txn_props(
             start_ts: TimeStamp,
             primary: &[u8],
             pessimistic: bool,
@@ -663,7 +635,7 @@ mod tests {
             }
         }
 
-        fn prewrite(&mut self, m: Mutation, pk: &[u8], start_ts: impl Into<TimeStamp>) {
+        pub fn prewrite(&mut self, m: Mutation, pk: &[u8], start_ts: impl Into<TimeStamp>) {
             let snap =
                 RegionSnapshot::<RocksSnapshot>::from_raw(self.db.c().clone(), self.region.clone());
             let start_ts = start_ts.into();
@@ -683,7 +655,7 @@ mod tests {
             self.write(txn.into_modifies());
         }
 
-        fn prewrite_pessimistic_lock(
+        pub fn prewrite_pessimistic_lock(
             &mut self,
             m: Mutation,
             pk: &[u8],
@@ -708,7 +680,7 @@ mod tests {
             self.write(txn.into_modifies());
         }
 
-        fn acquire_pessimistic_lock(
+        pub fn acquire_pessimistic_lock(
             &mut self,
             k: Key,
             pk: &[u8],
@@ -737,7 +709,7 @@ mod tests {
             self.write(txn.into_modifies());
         }
 
-        fn commit(
+        pub fn commit(
             &mut self,
             pk: &[u8],
             start_ts: impl Into<TimeStamp>,
@@ -753,7 +725,7 @@ mod tests {
             self.write(txn.into_modifies());
         }
 
-        fn rollback(&mut self, pk: &[u8], start_ts: impl Into<TimeStamp>) {
+        pub fn rollback(&mut self, pk: &[u8], start_ts: impl Into<TimeStamp>) {
             let snap =
                 RegionSnapshot::<RocksSnapshot>::from_raw(self.db.c().clone(), self.region.clone());
             let start_ts = start_ts.into();
@@ -771,7 +743,7 @@ mod tests {
             self.write(txn.into_modifies());
         }
 
-        fn gc(&mut self, pk: &[u8], safe_point: impl Into<TimeStamp> + Copy) {
+        pub fn gc(&mut self, pk: &[u8], safe_point: impl Into<TimeStamp> + Copy) {
             let cm = ConcurrencyManager::new(safe_point.into());
             loop {
                 let snap = RegionSnapshot::<RocksSnapshot>::from_raw(
@@ -789,7 +761,7 @@ mod tests {
             }
         }
 
-        fn write(&mut self, modifies: Vec<Modify>) {
+        pub fn write(&mut self, modifies: Vec<Modify>) {
             let db = &self.db;
             let mut wb = db.c().write_batch();
             for rev in modifies {
@@ -814,14 +786,14 @@ mod tests {
             wb.write().unwrap();
         }
 
-        fn flush(&mut self) {
+        pub fn flush(&mut self) {
             for cf in ALL_CFS {
                 let cf = engine_rocks::util::get_cf_handle(&self.db, cf).unwrap();
                 self.db.flush_cf(cf, true).unwrap();
             }
         }
 
-        fn compact(&mut self) {
+        pub fn compact(&mut self) {
             for cf in ALL_CFS {
                 let cf = engine_rocks::util::get_cf_handle(&self.db, cf).unwrap();
                 self.db.compact_range_cf(cf, None, None);
@@ -829,7 +801,7 @@ mod tests {
         }
     }
 
-    fn open_db(path: &str, with_properties: bool) -> Arc<DB> {
+    pub fn open_db(path: &str, with_properties: bool) -> Arc<DB> {
         let db_opts = DBOptions::new();
         let mut cf_opts = ColumnFamilyOptions::new();
         cf_opts.set_write_buffer_size(32 * 1024 * 1024);
@@ -846,7 +818,7 @@ mod tests {
         Arc::new(engine_rocks::raw_util::new_engine_opt(path, db_opts, cfs_opts).unwrap())
     }
 
-    fn make_region(id: u64, start_key: Vec<u8>, end_key: Vec<u8>) -> Region {
+    pub fn make_region(id: u64, start_key: Vec<u8>, end_key: Vec<u8>) -> Region {
         let mut peer = Peer::default();
         peer.set_id(id);
         peer.set_store_id(id);
@@ -856,55 +828,6 @@ mod tests {
         region.set_end_key(end_key);
         region.mut_peers().push(peer);
         region
-    }
-
-    fn get_mvcc_properties_and_check_gc(
-        db: Arc<DB>,
-        region: Region,
-        safe_point: impl Into<TimeStamp>,
-        need_gc: bool,
-    ) -> Option<MvccProperties> {
-        let safe_point = safe_point.into();
-
-        let start = keys::data_key(region.get_start_key());
-        let end = keys::data_end_key(region.get_end_key());
-        let props = db
-            .c()
-            .get_mvcc_properties_cf(CF_WRITE, safe_point, &start, &end);
-        if let Some(props) = props.as_ref() {
-            assert_eq!(check_need_gc(safe_point, 1.0, &props), need_gc);
-        }
-        props
-    }
-
-    #[test]
-    fn test_need_gc() {
-        let path = tempfile::Builder::new()
-            .prefix("test_storage_mvcc_reader")
-            .tempdir()
-            .unwrap();
-        let path = path.path().to_str().unwrap();
-        let region = make_region(1, vec![0], vec![10]);
-        test_without_properties(path, &region);
-        test_with_properties(path, &region);
-    }
-
-    fn test_without_properties(path: &str, region: &Region) {
-        let db = open_db(path, false);
-        let mut engine = RegionEngine::new(&db, &region);
-
-        // Put 2 keys.
-        engine.put(&[1], 1, 1);
-        engine.put(&[4], 2, 2);
-        assert!(
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, true).is_none()
-        );
-        engine.flush();
-        // After this flush, we have a SST file without properties.
-        // Without properties, we always need GC.
-        assert!(
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, true).is_none()
-        );
     }
 
     #[test]
@@ -1016,87 +939,6 @@ mod tests {
             key2
         );
         assert_eq!(iter.next().unwrap(), false);
-    }
-
-    fn test_with_properties(path: &str, region: &Region) {
-        let db = open_db(path, true);
-        let mut engine = RegionEngine::new(&db, &region);
-
-        // Put 2 keys.
-        engine.put(&[2], 3, 3);
-        engine.put(&[3], 4, 4);
-        engine.flush();
-        // After this flush, we have a SST file w/ properties, plus the SST
-        // file w/o properties from previous flush. We always need GC as
-        // long as we can't get properties from any SST files.
-        assert!(
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, true).is_none()
-        );
-        engine.compact();
-        // After this compact, the two SST files are compacted into a new
-        // SST file with properties. Now all SST files have properties and
-        // all keys have only one version, so we don't need gc.
-        let props =
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, false).unwrap();
-        assert_eq!(props.min_ts, 1.into());
-        assert_eq!(props.max_ts, 4.into());
-        assert_eq!(props.num_rows, 4);
-        assert_eq!(props.num_puts, 4);
-        assert_eq!(props.num_versions, 4);
-        assert_eq!(props.max_row_versions, 1);
-
-        // Put 2 more keys and delete them.
-        engine.put(&[5], 5, 5);
-        engine.put(&[6], 6, 6);
-        engine.delete(&[5], 7, 7);
-        engine.delete(&[6], 8, 8);
-        engine.flush();
-        // After this flush, keys 5,6 in the new SST file have more than one
-        // versions, so we need gc.
-        let props =
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, true).unwrap();
-        assert_eq!(props.min_ts, 1.into());
-        assert_eq!(props.max_ts, 8.into());
-        assert_eq!(props.num_rows, 6);
-        assert_eq!(props.num_puts, 6);
-        assert_eq!(props.num_versions, 8);
-        assert_eq!(props.max_row_versions, 2);
-        // But if the `safe_point` is older than all versions, we don't need gc too.
-        let props =
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 0, false).unwrap();
-        assert_eq!(props.min_ts, TimeStamp::max());
-        assert_eq!(props.max_ts, TimeStamp::zero());
-        assert_eq!(props.num_rows, 0);
-        assert_eq!(props.num_puts, 0);
-        assert_eq!(props.num_versions, 0);
-        assert_eq!(props.max_row_versions, 0);
-
-        // We gc the two deleted keys manually.
-        engine.gc(&[5], 10);
-        engine.gc(&[6], 10);
-        engine.compact();
-        // After this compact, all versions of keys 5,6 are deleted,
-        // no keys have more than one versions, so we don't need gc.
-        let props =
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, false).unwrap();
-        assert_eq!(props.min_ts, 1.into());
-        assert_eq!(props.max_ts, 4.into());
-        assert_eq!(props.num_rows, 4);
-        assert_eq!(props.num_puts, 4);
-        assert_eq!(props.num_versions, 4);
-        assert_eq!(props.max_row_versions, 1);
-
-        // A single lock version need gc.
-        engine.lock(&[7], 9, 9);
-        engine.flush();
-        let props =
-            get_mvcc_properties_and_check_gc(Arc::clone(&db), region.clone(), 10, true).unwrap();
-        assert_eq!(props.min_ts, 1.into());
-        assert_eq!(props.max_ts, 9.into());
-        assert_eq!(props.num_rows, 5);
-        assert_eq!(props.num_puts, 4);
-        assert_eq!(props.num_versions, 5);
-        assert_eq!(props.max_row_versions, 1);
     }
 
     #[test]
